Change mempool page realtime fee internals - using getmempoolentry now
The old approach was fetching every input for a tx and then calculating
the fee from the difference between inputs and outputs. This approach
was putting too much load on the server, an arm based device could not
keep up. Found out RPC has a getmempoolentry command, which shows the
fee directly. So now the RPC calls were reduced to 1 per zeroMQ event.

The only difference is that now fees from RPC now have to be multiplied
by 100 milion, so the bitcoinjs lib in the browser has no problem processing
it.

Resolves: #35
Obsoletes: #34 - there is no longer a need to make fee displaying obtional